+++ core-tests/jdk-8/spring-rest-openapi-v2-driver/src/main/java/com/foo/rest/examples/spring/endpointexclude/EndpointExcludeController.java
@@ -1,0 +1,21 @@
+package com.foo.rest.examples.spring.endpointexclude;
+
+import com.foo.rest.examples.spring.SpringController;
+import org.evomaster.client.java.controller.problem.ProblemInfo;
+import org.evomaster.client.java.controller.problem.RestProblem;
+
+public class EndpointExcludeController extends SpringController {
+
+    public EndpointExcludeController(){
+        super(EndpointExcludeApplication.class);
+    }
+
+    @Override
+    public ProblemInfo getProblemInfo() {
+
+        return new RestProblem(
+                "http://localhost:" + getSutPort() + "/v2/api-docs",
+                null
+        );
+    }
+}