+++ core-tests/jdk-8/spring-rest-openapi-v2-driver/src/main/java/com/foo/rest/examples/spring/valid/ValidController.java
@@ -1,0 +1,14 @@
+package com.foo.rest.examples.spring.valid;
+
+import com.foo.rest.examples.spring.SpringController;
+
+/**
+ * Created by arcuri82 on 07-Nov-18.
+ */
+public class ValidController extends SpringController {
+
+    public ValidController() {
+        super(ValidApplication.class);
+    }
+
+}